start to use autotools
Using non-recursive automake now, needs automake 1.14 or above if
changing Makefile.am or one of the local.mk files, as older versions may
not support %C% and %D% notations.

AUTHORS, README and other standard files are added, please consider
updating them when necessary.

make distcheck is passed before this commit, and the source is ready for
distribution. By the way, all this initially was for adding the
automatic detection or user selection of event devices. I have written a
python script that shows a nice curses interface for device selection,
and next step is adding this mechanism to autoconf.