Adjustment on compiling flags
Quite many codes in the project needs c99 standards, such as variadic
arguments in macros, but the Makefile used to adopt iso9899:1990 (c90).

configure now also checks if the compiling flags are available in
detected compiler. If gcc is not used, other compilers are still
possible to support these features by their own extensions. In that
case, not compiling flags is explicitly given.