try to fix problem with fb's odd line_length
For example, on a framebuffer with a picture size of 1366x768, the line
length(bytes per line in fb) may not be 1366*color_depth/CHAR_BIT, but
something larger than this. I guess this is for alignment on the
hardware level.

So the use of screen width in graph_lower functions for determining the
line length should be substituted by global_fix_screen_info.line_length.

But so far the problem is not yet solved.