fixes and format adjustment
Checkbox widget can almost work properly now. Sample program
multi_windows.c can now show a checkbox on the top left corner.

But the initialization with default style need to be checked again for
other widgets, as many of them just ignores the default style and leave
the widget uninitialized if the user-defined style file is not valid.