testing code for odd object trees
Windows and other widgets are organized as a tree of objects, but not so
simple. It seems that the tree is not well formed, which causes problems
when a window becomes a child of another window. For example, the
function widget_absolute_area() can't work properly.

This branch is an attempt to rewrite a large part of code about this.