tree structure in widgets
Trying to build a real tree structure in widgets, each widget saving a
list of children, and recording its parent in parent field in OBJECT.

This time we just did some simplification in the code, utilizing more
code in the widget parent class. Code in the application library should
be modified next.

The change in this commit is NOT TESTED.