fix tab page header inconsistency
The header buttons didn't appear pressed at program start, and when
repainted they also recovered to the released state because we didn't
catch the repaint signal.

The solution for the first problem (at program start) is not a very good
one, which simply sets the first page header pressed. There may be a
better one.