skeleton of new header for tab widget
Trying to make the tab header like the sheet selection tab in
LibreOffice Calc. Headers of all the tab pages will make up a whole
tab_header widget, which will check cursor position on mouse press and
select the tab header under cursor.

For this new widget, the button in tab_page widget is removed, and now
tab_page is a derived widget from panel. Panel is modified for being
inherited beforehand in master branch.

As the button in tab_page is removed, it is temporarily not supported to
change font for individual header of a tab page. This may be added back
by changing the page_head in tab_page widget into a label.