tab_header development commit
tab_page is rolling back to using macro for struct inheriting.

panel is using a default width and height of 400. This should be
meaningless, but not wrong. It's not hard-coding, as an rc file can
override this if desired.

tab is going to be like a "shell" widget, as most of the job including
transfer of tab pages are given to tab_header.