adapt code from fbida for framebuffer init
It now supports switching between ttys. Only the tty it's opened in will
show graphics. So if the guest programs are run from another tty (which
is true for every sample program now), you have to move back to the
original tty to view them.

But there are problems: if tty is changed during wm's load time, the
graphics will leak to the new tty, and control remains in the new one
until you switch back; tty has to be manually reset after server quitted
normally.

The first problem above is the same in fbida itself, except that the
control doesn't leak.